refactor: StudyId Long으로 타입 수정

+++ android/app/src/main/java/com/created/team201/data/remote/response/HomeStudyResponseDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class HomeStudyResponseDto(
     @SerialName("id")
-    val studyId: Int,
+    val studyId: Long,
     @SerialName("name")
     val studyName: String,
     @SerialName("todoContent")
